refactor: factor out exported types from chain-service
