+++ packages/server-wallet/src/chain-service/index.ts
@@ -1,2 +1,3 @@
 export * from './chain-service';
 export * from './mock-chain-service';
+export * from './types';
